DB 멤버 테이블, 멤버객체 생성, 회원가입 view단 , back단 로직 구현

+++ src/main/java/com/ssh/example/demo/controller/UsrArticleController.java
@@ -35,7 +35,7 @@ public class UsrArticleController {
 
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
 
-		int id = writeArticleRd.getData1();
+		int id = (Integer)writeArticleRd.getData1();
 
 		return ResultData.from(writeArticleRd.getResultCode(),writeArticleRd.getMsg(), id);
